+++ src/typecheck.cc
@@ -35,7 +35,7 @@ C_type TypeChecker::visit(const Literal &e) {
   return std::visit(
       [](const auto &arg) {
         using T = decay_t<decltype(arg)>;
-        if constexpr (is_same_v<T, int>)
+        if constexpr (is_same_v<T, long>)
           return C_type::C_int;
         else if constexpr (is_same_v<T, bool>)
           return C_type::C_bool;
